Branch snapshot for CL 3474273
[CL 3474273 in Dev-ARKit branch]

+++ Engine/Source/Runtime/Engine/Private/UnrealEngine.cpp
@@ -7354,7 +7354,7 @@ void UEngine::PerformanceCapture(UWorld* World, const FString& MapName, const FS
 		UConsole* ViewportConsole = (GEngine->GameViewport != nullptr) ? GEngine->GameViewport->ViewportConsole : nullptr;
 		FConsoleOutputDevice StrOut(ViewportConsole);
 
-		StrOut.Logf(TEXT("  frame:%d %s"), GFrameCounter, *ScreenshotName);
+		StrOut.Logf(TEXT("  frame:%llu %s"), (uint64)GFrameCounter, *ScreenshotName);
 	}
 
 	const bool bShowUI = false;
